requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
scikit-learn==1.4.1.post1
pythainlp==4.0.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0